Rename compose to imperative
